refactoring movie list
