+++ src/containers/FilmDescription.js
@@ -1,9 +1,7 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import { MovieCard } from '../components/MovieCard';
 import { MovieList } from '../components/MovieList';
 import { MovieCardDescription } from '../components/MovieCardDescription';
-import * as css from '../styles/start-page.css';
 import * as cssList from '../styles/movie-list.css';
 
 
@@ -13,26 +11,20 @@ export class FilmDescription extends React.Component {
   }
 
   componentWillReceiveProps(nextProps) {
+    console.log("I am inside");
     if(nextProps.match.params.filmName != this.props.match.params.filmName) {
+      console.log("I am here");
       this.props.loadMovieInfo(nextProps.match.params.filmName);
     }
   }
 
   render() {
-    const relatedMovies = this.props.movies.map((movie, index) => {
-      return <MovieCard
-        title={movie.show_title}
-        year={movie.release_year}
-        poster={movie.poster}
-        category={movie.category}
-        key={index}
-      />
-    });
-
     const movieListBar =
       <div className={cssList.resultsContainer + " container"}>
-        {this.props.currentMovie.director ? <div>Films by {this.props.currentMovie.director}</div> : ""}
+        <div>Films by {this.props.currentMovie.director}</div>
       </div>;
+
+    const emptyMovieListBar = <div className={cssList.resultsContainer + " container"}></div>;
 
     return (
       <div>
@@ -47,12 +39,9 @@ export class FilmDescription extends React.Component {
             </div>
           </header>
           <main>
-            <MovieList>
+            <MovieList movies={this.props.movies}>
               <div className={cssList.resultsBar}>
-                { movieListBar }
-              </div>
-              <div className="flex container relative">
-                { relatedMovies }
+                { this.props.currentMovie.length ? movieListBar : emptyMovieListBar}
               </div>
             </MovieList>
           </main>
